requirements.txt
streamlit==1.36.0
pydantic>=2.7
requests>=2.31
google-cloud-firestore>=2.16.0
google-auth>=2.29.0
python-dotenv>=1.0.1
altair>=5.0
pandas>=2.1
